backend/requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.12.0
asgiref==3.11.0
async-timeout==5.0.1
attrs==25.4.0
billiard==4.2.4
cachetools==6.2.4
celery==5.6.1
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cryptography==46.0.3
distro==1.9.0
Django==5.2.9
django-cors-headers==4.9.0
django-filter==25.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
exceptiongroup==1.3.1
faiss-cpu==1.13.2
filelock==3.20.0
fsspec==2025.12.0
google-ai-generativelanguage==0.9.0
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.45.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.3
google-genai==1.56.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.76.0
gunicorn==23.0.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
inflection==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.2
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
oauthlib==3.3.1
openai==2.14.0
packaging==25.0
pillow==12.1.0
prompt_toolkit==3.0.52
proto-plus==1.27.0
protobuf==6.33.2
psycopg2-binary==2.9.11
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
PyJWT==2.10.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-http-client==3.3.7
PyYAML==6.0.3
redis==7.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
sendgrid==6.12.5
sentence-transformers==5.2.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.9.1+cpu
tqdm==4.67.1
transformers==4.57.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.2
vine==5.1.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.4
